moving qr-code logic to untested package
The qr-code logic is not used for most tests, speeding them up.  This makes running the handler tests 1000 times take 7.29 seconds (from 21.91 seconds, a 235% speedup).
Adding server integration test that generates board with a qr code.  The test also checks to ensure the id is on the board rendering (it is in the free space).  Thistest can be skipped by running tests with the '--short' flag (the test takes about 2 seconds for 1000 runs).